Refactor GuessWho and update println
